experiments/universal_model_tester/requirements.txt
litellm>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0